backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
openai==1.3.7
pydantic==2.5.0
pydantic-settings==2.1.0
